Adding Omnidirectional Shadow Maps

+++ DirectionalLight.h
@@ -4,7 +4,7 @@ class DirectionalLight : public Light
 {
 public:
 	DirectionalLight();
-	DirectionalLight(GLfloat shadowWidth, GLfloat shadowHeight, 
+	DirectionalLight(GLuint shadowWidth, GLuint shadowHeight,
 					GLfloat red, GLfloat green, GLfloat blue,
 					GLfloat aIntensity, GLfloat dIntensity,
 					GLfloat xDir, GLfloat yDir, GLfloat zDir);
